Fix performance bug in allocSlice
Function allocSlice could return slices with a smaller capacity than what it was required if these slices are pulled from the last pool.

This wasn't an issue in practice as clauses properly resized their underlying slices. However, this bug led to doing more allocations than necessary (e.g. if the returned slice has a much smaller capacity than what was requested).

+++ internal/sat/clauses.go
@@ -76,7 +76,7 @@ func NewClause(s *Solver, tmpLiterals []Literal, learnt bool) (*Clause, bool) {
 		c := &Clause{}
 		c.learnt = learnt
 
-		c.sliceRef = allocSlice(uint(size))
+		c.sliceRef = allocSlice(size) // reuse already allocated slices
 		c.literals = *c.sliceRef
 		c.literals = append(c.literals, tmpLiterals...)
 
@@ -108,7 +108,7 @@ func (c *Clause) Delete(s *Solver) {
 	s.Unwatch(c, c.literals[1].Opposite())
 
 	*c.sliceRef = c.literals
-	freeSlice(c.sliceRef)
+	freeSlice(c.sliceRef) // give the slice back for other clauses to use
 }
 
 func (c *Clause) Simplify(s *Solver) bool {
